--- alembic/versions/13b45e8e36f0_new_migration.py
@@ -1,28 +1,0 @@
-"""New migration
-
-Revision ID: 13b45e8e36f0
-Revises: add69e681ebe
-Create Date: 2025-09-29 17:52:47.453978
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '13b45e8e36f0'
-down_revision = 'add69e681ebe'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('hashed_refresh_token', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'hashed_refresh_token')
-    # ### end Alembic commands ###
