--- alembic/versions/6a065647d9f6_new_migration.py
@@ -1,293 +1,0 @@
-"""New migration
-
-Revision ID: 6a065647d9f6
-Revises: b91056978982
-Create Date: 2025-09-27 17:23:31.578815
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '6a065647d9f6'
-down_revision = 'b91056978982'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('investment',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('payment',
-    sa.Column('recorder_id', sa.UUID(), nullable=False),
-    sa.Column('amount', sa.BigInteger(), nullable=False),
-    sa.Column('payment_method', sa.Enum('CASH', 'CARD_TRANSACTION', 'POS_MACHINE', name='paymentmethod', native_enum=False), nullable=False),
-    sa.Column('direction', sa.Enum('INCOMING', 'OUTGOING', 'INTERNAL_TRANSFER', name='paymentdirection', native_enum=False), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('photo_holder_id', sa.UUID(), nullable=True),
-    sa.Column('status', sa.Enum('DRAFT', 'APPROVED_BY_USER', 'APPROVED_BY_ADMIN', name='approvalstatus', native_enum=False), nullable=False),
-    sa.Column('investment_id', sa.UUID(), nullable=True),
-    sa.Column('transaction_id', sa.UUID(), nullable=True),
-    sa.Column('account_ledger_id', sa.UUID(), nullable=True),
-    sa.Column('saved_bank_account_id', sa.UUID(), nullable=True),
-    sa.Column('contact_id', sa.UUID(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['account_ledger_id'], ['account_ledger.id'], ondelete='SET NULL'),
-    sa.ForeignKeyConstraint(['contact_id'], ['contact.id'], ondelete='RESTRICT'),
-    sa.ForeignKeyConstraint(['investment_id'], ['investment.id'], ondelete='SET NULL'),
-    sa.ForeignKeyConstraint(['photo_holder_id'], ['user.id'], ),
-    sa.ForeignKeyConstraint(['recorder_id'], ['user.id'], ),
-    sa.ForeignKeyConstraint(['saved_bank_account_id'], ['saved_bank_account.id'], ondelete='SET NULL'),
-    sa.ForeignKeyConstraint(['transaction_id'], ['transaction.id'], ondelete='SET NULL'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_payment_account_ledger_id'), 'payment', ['account_ledger_id'], unique=False)
-    op.create_index(op.f('ix_payment_contact_id'), 'payment', ['contact_id'], unique=False)
-    op.create_index(op.f('ix_payment_investment_id'), 'payment', ['investment_id'], unique=False)
-    op.create_index(op.f('ix_payment_recorder_id'), 'payment', ['recorder_id'], unique=False)
-    op.create_index(op.f('ix_payment_saved_bank_account_id'), 'payment', ['saved_bank_account_id'], unique=False)
-    op.create_index(op.f('ix_payment_status'), 'payment', ['status'], unique=False)
-    op.create_index(op.f('ix_payment_transaction_id'), 'payment', ['transaction_id'], unique=False)
-    op.alter_column('account_ledger', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('account_ledger', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('audit_log', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('audit_log', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('contact', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('contact', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('inventory', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('inventory', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item_financial_profile', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item_financial_profile', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('permission', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('permission', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('saved_bank_account', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('saved_bank_account', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction_item', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction_item', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('user', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('user', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.DateTime(timezone=True),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('user', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction_item', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction_item', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('transaction', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('saved_bank_account', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('saved_bank_account', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('permission', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('permission', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item_financial_profile', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item_financial_profile', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('item', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('inventory', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('inventory', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('contact', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('contact', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('audit_log', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('audit_log', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('account_ledger', 'updated_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('account_ledger', 'created_at',
-               existing_type=sa.DateTime(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True,
-               existing_server_default=sa.text('now()'))
-    op.drop_index(op.f('ix_payment_transaction_id'), table_name='payment')
-    op.drop_index(op.f('ix_payment_status'), table_name='payment')
-    op.drop_index(op.f('ix_payment_saved_bank_account_id'), table_name='payment')
-    op.drop_index(op.f('ix_payment_recorder_id'), table_name='payment')
-    op.drop_index(op.f('ix_payment_investment_id'), table_name='payment')
-    op.drop_index(op.f('ix_payment_contact_id'), table_name='payment')
-    op.drop_index(op.f('ix_payment_account_ledger_id'), table_name='payment')
-    op.drop_table('payment')
-    op.drop_table('investment')
-    # ### end Alembic commands ###
