--- alembic/versions/a346c0bdd816_new_migration.py
@@ -1,77 +1,0 @@
-"""New migration
-
-Revision ID: a346c0bdd816
-Revises: 4374d73ff647
-Create Date: 2025-09-20 20:22:52.448124
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a346c0bdd816'
-down_revision = '4374d73ff647'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('transaction',
-    sa.Column('recorder_id', sa.UUID(), nullable=False),
-    sa.Column('contact_id', sa.UUID(), nullable=False),
-    sa.Column('note', sa.Text(), nullable=True),
-    sa.Column('status', sa.Enum('DRAFT', 'APPROVED_BY_USER', 'APPROVED_BY_ADMIN', name='transactionstatus', native_enum=False), nullable=False),
-    sa.Column('discount', sa.BigInteger(), nullable=False),
-    sa.Column('total_price', sa.BigInteger(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['contact_id'], ['contact.id'], ),
-    sa.ForeignKeyConstraint(['recorder_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_transaction_contact_id'), 'transaction', ['contact_id'], unique=False)
-    op.create_index(op.f('ix_transaction_recorder_id'), 'transaction', ['recorder_id'], unique=False)
-    op.create_index(op.f('ix_transaction_status'), 'transaction', ['status'], unique=False)
-    op.create_table('transaction_item',
-    sa.Column('transaction_id', sa.UUID(), nullable=False),
-    sa.Column('item_id', sa.UUID(), nullable=False),
-    sa.Column('transaction_type', sa.Enum('BUY', 'SELL', name='transactiontype', native_enum=False), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('weight_count', sa.Numeric(precision=12, scale=2), nullable=False),
-    sa.Column('unit_price', sa.BigInteger(), nullable=False),
-    sa.Column('total_price', sa.BigInteger(), nullable=False),
-    sa.Column('karat', sa.Numeric(precision=10, scale=2), nullable=True),
-    sa.Column('ojrat', sa.Numeric(precision=10, scale=2), nullable=True),
-    sa.Column('profit', sa.Numeric(precision=10, scale=2), nullable=True),
-    sa.Column('tax', sa.Numeric(precision=10, scale=2), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['item_id'], ['item.id'], ),
-    sa.ForeignKeyConstraint(['transaction_id'], ['transaction.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_transaction_item_item_id'), 'transaction_item', ['item_id'], unique=False)
-    op.create_index(op.f('ix_transaction_item_title'), 'transaction_item', ['title'], unique=False)
-    op.create_index(op.f('ix_transaction_item_transaction_id'), 'transaction_item', ['transaction_id'], unique=False)
-    op.add_column('inventory', sa.Column('transaction_id', sa.UUID(), nullable=True))
-    op.create_foreign_key(None, 'inventory', 'transaction', ['transaction_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'inventory', type_='foreignkey')
-    op.drop_column('inventory', 'transaction_id')
-    op.drop_index(op.f('ix_transaction_item_transaction_id'), table_name='transaction_item')
-    op.drop_index(op.f('ix_transaction_item_title'), table_name='transaction_item')
-    op.drop_index(op.f('ix_transaction_item_item_id'), table_name='transaction_item')
-    op.drop_table('transaction_item')
-    op.drop_index(op.f('ix_transaction_status'), table_name='transaction')
-    op.drop_index(op.f('ix_transaction_recorder_id'), table_name='transaction')
-    op.drop_index(op.f('ix_transaction_contact_id'), table_name='transaction')
-    op.drop_table('transaction')
-    # ### end Alembic commands ###
