--- alembic/versions/a771eba8b681_new_migration.py
@@ -1,30 +1,0 @@
-"""New migration
-
-Revision ID: a771eba8b681
-Revises: 6a065647d9f6
-Create Date: 2025-09-27 18:04:27.794570
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a771eba8b681'
-down_revision = '6a065647d9f6'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('inventory', sa.Column('payment_id', sa.UUID(), nullable=True))
-    op.create_foreign_key(None, 'inventory', 'payment', ['payment_id'], ['id'], ondelete='SET NULL')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'inventory', type_='foreignkey')
-    op.drop_column('inventory', 'payment_id')
-    # ### end Alembic commands ###
