--- alembic/versions/add69e681ebe_new_migration.py
@@ -1,28 +1,0 @@
-"""New migration
-
-Revision ID: add69e681ebe
-Revises: a771eba8b681
-Create Date: 2025-09-28 14:19:41.593144
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'add69e681ebe'
-down_revision = 'a771eba8b681'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
