--- alembic/versions/b91056978982_new_migration.py
@@ -1,50 +1,0 @@
-"""New migration
-
-Revision ID: b91056978982
-Revises: b95fef3331cd
-Create Date: 2025-09-26 18:24:29.268367
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'b91056978982'
-down_revision = 'b95fef3331cd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('account_ledger',
-    sa.Column('contact_id', sa.UUID(), nullable=False),
-    sa.Column('transaction_id', sa.UUID(), nullable=True),
-    sa.Column('debt', sa.BigInteger(), nullable=False, comment='Money that the shopkeeper owes to the customer (the reverse is not currently supported)'),
-    sa.Column('deadline', sa.DateTime(), nullable=True),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('card_number', sa.String(), nullable=True),
-    sa.Column('bank_name', sa.String(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['contact_id'], ['contact.id'], ondelete='RESTRICT'),
-    sa.ForeignKeyConstraint(['transaction_id'], ['transaction.id'], ondelete='SET NULL'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_account_ledger_bank_name'), 'account_ledger', ['bank_name'], unique=False)
-    op.create_index(op.f('ix_account_ledger_contact_id'), 'account_ledger', ['contact_id'], unique=False)
-    op.create_index(op.f('ix_account_ledger_deadline'), 'account_ledger', ['deadline'], unique=False)
-    op.create_index(op.f('ix_account_ledger_transaction_id'), 'account_ledger', ['transaction_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_account_ledger_transaction_id'), table_name='account_ledger')
-    op.drop_index(op.f('ix_account_ledger_deadline'), table_name='account_ledger')
-    op.drop_index(op.f('ix_account_ledger_contact_id'), table_name='account_ledger')
-    op.drop_index(op.f('ix_account_ledger_bank_name'), table_name='account_ledger')
-    op.drop_table('account_ledger')
-    # ### end Alembic commands ###
