--- alembic/versions/b95fef3331cd_new_migration.py
@@ -1,30 +1,0 @@
-"""New migration
-
-Revision ID: b95fef3331cd
-Revises: a346c0bdd816
-Create Date: 2025-09-23 19:48:06.798752
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'b95fef3331cd'
-down_revision = 'a346c0bdd816'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f('inventory_transaction_id_fkey'), 'inventory', type_='foreignkey')
-    op.create_foreign_key(None, 'inventory', 'transaction', ['transaction_id'], ['id'], ondelete='SET NULL')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'inventory', type_='foreignkey')
-    op.create_foreign_key(op.f('inventory_transaction_id_fkey'), 'inventory', 'transaction', ['transaction_id'], ['id'])
-    # ### end Alembic commands ###
